+++ addon/components/object-list-view.js
@@ -290,6 +290,15 @@ export default FlexberryBaseComponent.extend(
   showDeleteButtonInRow: false,
 
   /**
+    Flag indicates whether to show edit button in first column of every row.
+
+    @property showEditButtonInRow
+    @type Boolean
+    @default false
+  */
+  showEditButtonInRow: false,
+
+  /**
     Flag indicates whether to not use userSetting from backend
     @property notUseUserSettings
     @type Boolean
@@ -308,10 +317,11 @@ export default FlexberryBaseComponent.extend(
     'showAsteriskInRow',
     'showCheckBoxInRow',
     'showDeleteButtonInRow',
+    'showEditButtonInRow',
     'modelProjection',
     function() {
       if (this.get('modelProjection')) {
-        return this.get('showAsteriskInRow') || this.get('showCheckBoxInRow') || this.get('showDeleteButtonInRow');
+        return this.get('showAsteriskInRow') || this.get('showCheckBoxInRow') || this.get('showDeleteButtonInRow') || this.get('showEditButtonInRow');
       } else {
         return false;
       }
@@ -801,7 +811,7 @@ export default FlexberryBaseComponent.extend(
         }
       }
 
-      if (this.rowClickable) {
+      if (this.rowClickable || params.rowEdit) {
         let recordKey = recordWithKey && recordWithKey.key;
         let recordData = recordWithKey && recordWithKey.data;
 
@@ -1230,8 +1240,11 @@ export default FlexberryBaseComponent.extend(
           if (currentPropertyName === 'OlvRowToolbar') {
             let checkbox = this.get('showCheckBoxInRow');
             let delButton = this.get('showDeleteButtonInRow');
+            let editButton = this.get('showEditButtonInRow');
 
-            setting.width = (checkbox && delButton ? 100 : delButton ? 70 : 65) - padding;
+            // TODO: Probably need to replace this.
+            setting.width = (checkbox && delButton && editButton ? 130 : (checkbox && delButton) || (checkbox && editButton) ||
+            (delButton && editButton) ? 100 : checkbox || delButton || editButton ? 70 : 65) - padding;
           }
         }